+++ app/dependencies/inmemoryDebugRuntimeClasspath.txt
@@ -1,4 +1,5 @@
 :app-flavor-inmemory
+:compose-ui
 :mock-data
 :models
 :networking
@@ -19,24 +20,23 @@ androidx.collection:collection:1.1.0
 androidx.compose.animation:animation-core:1.2.0
 androidx.compose.animation:animation:1.2.0
 androidx.compose.compiler:compiler:1.3.0
-androidx.compose.foundation:foundation-layout:1.2.0
-androidx.compose.foundation:foundation:1.2.0
-androidx.compose.material:material-icons-core:1.2.0
+androidx.compose.foundation:foundation-layout:1.2.1
+androidx.compose.foundation:foundation:1.2.1
+androidx.compose.material:material-icons-core:1.2.1
 androidx.compose.material:material-icons-extended:1.2.0
-androidx.compose.material:material-ripple:1.2.0
-androidx.compose.material:material:1.2.0
-androidx.compose.runtime:runtime-livedata:1.2.0
-androidx.compose.runtime:runtime-saveable:1.2.0
-androidx.compose.runtime:runtime:1.2.0
-androidx.compose.ui:ui-geometry:1.2.0
-androidx.compose.ui:ui-graphics:1.2.0
-androidx.compose.ui:ui-text:1.2.0
+androidx.compose.material:material-ripple:1.2.1
+androidx.compose.material:material:1.2.1
+androidx.compose.runtime:runtime-saveable:1.2.1
+androidx.compose.runtime:runtime:1.2.1
+androidx.compose.ui:ui-geometry:1.2.1
+androidx.compose.ui:ui-graphics:1.2.1
+androidx.compose.ui:ui-text:1.2.1
 androidx.compose.ui:ui-tooling-data:1.2.0
 androidx.compose.ui:ui-tooling-preview:1.2.0
 androidx.compose.ui:ui-tooling:1.2.0
-androidx.compose.ui:ui-unit:1.2.0
-androidx.compose.ui:ui-util:1.2.0
-androidx.compose.ui:ui:1.2.0
+androidx.compose.ui:ui-unit:1.2.1
+androidx.compose.ui:ui-util:1.2.1
+androidx.compose.ui:ui:1.2.1
 androidx.concurrent:concurrent-futures:1.0.0
 androidx.constraintlayout:constraintlayout-solver:2.0.1
 androidx.constraintlayout:constraintlayout:2.0.1
@@ -58,7 +58,7 @@ androidx.legacy:legacy-support-core-utils:1.0.0
 androidx.lifecycle:lifecycle-common-java8:2.4.1
 androidx.lifecycle:lifecycle-common:2.4.1
 androidx.lifecycle:lifecycle-livedata-core:2.3.1
-androidx.lifecycle:lifecycle-livedata:2.2.0
+androidx.lifecycle:lifecycle-livedata:2.0.0
 androidx.lifecycle:lifecycle-process:2.4.0
 androidx.lifecycle:lifecycle-runtime-ktx:2.4.1
 androidx.lifecycle:lifecycle-runtime:2.4.1
@@ -123,6 +123,9 @@ io.ktor:ktor-websocket-serialization-jvm:2.1.0
 io.ktor:ktor-websocket-serialization:2.1.0
 io.ktor:ktor-websockets-jvm:2.1.0
 io.ktor:ktor-websockets:2.1.0
+org.jetbrains.compose.foundation:foundation:1.2.0-alpha01-dev770
+org.jetbrains.compose.material:material:1.2.0-alpha01-dev770
+org.jetbrains.compose.ui:ui:1.2.0-alpha01-dev770
 org.jetbrains.kotlin:kotlin-android-extensions-runtime:1.5.10
 org.jetbrains.kotlin:kotlin-parcelize-runtime:1.5.10
 org.jetbrains.kotlin:kotlin-stdlib-common:1.7.10
