+++ app/src/main/java/com/handstandsam/shoppingapp/features/checkout/CheckoutActivity.kt
@@ -4,36 +4,14 @@ import android.os.Bundle
 import android.view.Gravity
 import android.widget.Toast
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.size
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import com.handstandsam.shoppingapp.LoggedInActivity
-import com.handstandsam.shoppingapp.compose.ShoppingAppImageLoader
+import com.handstandsam.shoppingapp.compose.AndroidShoppingAppImageLoader
 import com.handstandsam.shoppingapp.compose.ShoppingCartScreen
-import com.handstandsam.shoppingapp.compose.loadNetworkImage
 import com.handstandsam.shoppingapp.features.home.AndroidShoppingCartViewModel
-import com.skydoves.landscapist.coil.CoilImage
-
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-
-
-class AndroidShoppingAppImageLoader : ShoppingAppImageLoader {
-    @Composable
-    override fun loadImage(modifier: Modifier, image: String) {
-        return CoilImage(
-            modifier = modifier,
-            imageModel = image,
-            // Crop, Fit, Inside, FillHeight, FillWidth, None
-            contentScale = ContentScale.Crop,
-            // shows an image with a circular revealed animation.
-            circularRevealedEnabled = true
-        )
-    }
-}
 
 class CheckoutActivity : LoggedInActivity() {
 
@@ -76,7 +54,8 @@ class CheckoutActivity : LoggedInActivity() {
                     )
                 },
                 logoutClicked = { logout() },
-                homeUpClicked = { onBackPressed() }
+                homeUpClicked = { onBackPressed() },
+                shoppingAppImageLoader = AndroidShoppingAppImageLoader()
             )
         }
     }
