+++ compose-ui/build.gradle.kts
@@ -1,5 +1,4 @@
 import org.jetbrains.compose.compose
-import org.jetbrains.compose.desktop.application.dsl.TargetFormat
 
 plugins {
     kotlin("multiplatform") // kotlin("jvm") doesn't work well in IDEA/AndroidStudio (https://github.com/JetBrains/compose-jb/issues/22)
@@ -10,18 +9,17 @@ kotlin {
     jvm {
         withJava()
     }
-    sourceSets {
-        named("jvmMain") {
-            dependencies {
-                compileOnly(compose.ui)
-                compileOnly(compose.foundation)
-                compileOnly(compose.material)
-                implementation(libs.kotlin.coroutines)
-                implementation(project(":models"))
-                implementation(project(":mock-data"))
-                implementation(project(":shopping-cart"))
-                implementation(project(":networking"))
-            }
+    val jvmMain by sourceSets.getting {
+        dependencies {
+            implementation(compose.ui)
+            implementation(compose.foundation)
+            implementation(compose.material)
+
+            implementation(project(":models"))
+            implementation(project(":mock-data"))
+            implementation(project(":shopping-cart"))
+            implementation(project(":networking"))
+            implementation(libs.kotlin.coroutines)
         }
     }
 }
